requirements.txt
cassandra-driver==3.29.0
tabulate==0.9.0
time-uuid==0.3.0